Moving location of 'types' module.
