+++ tripleo/utils/types.py
